+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### ✨ 新增功能
+
+- 新增领地 **名称**、**描述** 检查 @engsr6982
+
 ### 🐛 问题修复
 
 - 修复领地最小间距计算错误 @engsr6982
@@ -27,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 新增 `LandRefundFailedEvent` 事件 @engsr6982
 - 新增 `PlayerRequestCreateLandEvent` 事件 @engsr6982
 - 移除 `PlayerAskCreateLandEvent` 事件 @engsr6982
+- 新增 `PlayerChangeLandNameBeforeEvent`、`PlayerChangeLandNameAfterEvent` 事件 @engsr6982
 
 ## [0.17.1] - 2026-01-28